@@ -1,17 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="2.2.0" />
-      <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0" />
-      <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="6.0.9" />
-      <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.23.1" />
-      <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.23.1" />
-    </ItemGroup>
-
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="2.2.0"/>
+    <PackageReference Include="Microsoft.AspNetCore.Identity" Version="2.2.0"/>
+    <PackageReference Include="Microsoft.Extensions.Identity.Stores" Version="6.0.9"/>
+    <PackageReference Include="Microsoft.IdentityModel.Tokens" Version="6.23.1"/>
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.23.1"/>
+  </ItemGroup>
+  <PropertyGroup>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <TargetFramework>net7.0</TargetFramework>
+  </PropertyGroup>
 </Project>
