@@ -1,21 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Microsoft.AspNetCore.Hosting.Server.Abstractions" Version="2.2.0" />
-      <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5" />
-      <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.4.0" />
-      <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.4.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\TwojUrlop.Common\TwojUrlop.Common.csproj" />
-      <ProjectReference Include="..\TwojUrlop.DataAccess\TwojUrlop.DataAccess.csproj" />
-    </ItemGroup>
-
+  <ItemGroup>
+    <PackageReference Include="FastExpressionCompiler" Version="3.3.3"/>
+    <PackageReference Include="Mapster" Version="7.3.0"/>
+    <PackageReference Include="Microsoft.AspNetCore.Hosting.Server.Abstractions" Version="2.2.0"/>
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Core" Version="2.2.5"/>
+    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.4.0"/>
+    <PackageReference Include="Swashbuckle.AspNetCore.SwaggerUI" Version="6.4.0"/>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\TwojUrlop.Common\TwojUrlop.Common.csproj"/>
+    <ProjectReference Include="..\TwojUrlop.DataAccess\TwojUrlop.DataAccess.csproj"/>
+  </ItemGroup>
+  <PropertyGroup>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <TargetFramework>net7.0</TargetFramework>
+  </PropertyGroup>
 </Project>
