@@ -1,26 +1,24 @@
 using Microsoft.AspNetCore.Builder;
 
-namespace TwojUrlop.DependencyInjection.Extensions
+namespace TwojUrlop.DependencyInjection.Extensions;
+public static class ApplicationBuilderExtensions
 {
-    public static class ApplicationBuilderExtensions
+    public static void ConfigureCommonPipeline(this IApplicationBuilder app, bool isDevelopment)
     {
-        public static void ConfigureCommonPipeline(this IApplicationBuilder app, bool isDevelopment)
+        if (isDevelopment)
         {
-            if (isDevelopment)
-            {
-                app.UseDeveloperExceptionPage();
-                app.UseSwagger();
-            }
-            if(!isDevelopment)
-                app.UseHttpsRedirection();
-
-            app.UseRouting();
-            app.UseAuthentication();
-            app.UseAuthorization();
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
+            app.UseDeveloperExceptionPage();
+            app.UseSwagger();
         }
+        if (!isDevelopment)
+            app.UseHttpsRedirection();
+
+        app.UseRouting();
+        app.UseAuthentication();
+        app.UseAuthorization();
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllers();
+        });
     }
 }
