@@ -1,17 +1,15 @@
 using Microsoft.AspNetCore.Builder;
 
-namespace TwojUrlop.DependencyInjection.Extensions
+namespace TwojUrlop.DependencyInjection.Extensions;
+public static class CorsConfigExtensions
 {
-    public static class CorsConfigExtensions
+    public static void UseCorsConfig(this IApplicationBuilder app)
     {
-        public static void UseCorsConfig(this IApplicationBuilder app)
+        app.UseCors(options =>
         {
-            app.UseCors(options =>
-            {
-                options.AllowCredentials();
-                options.AllowAnyMethod();
-                options.AllowAnyHeader();
-            });
-        }
+            options.AllowCredentials();
+            options.AllowAnyMethod();
+            options.AllowAnyHeader();
+        });
     }
 }
