@@ -1,15 +1,14 @@
-﻿using System;
+﻿using TwojUrlop.Domain.Authorization.Commands;
+using TwojUrlop.DomainModel.Authorization.Commands.SignUp;
+
 namespace TwojUrlop.Extensions
 {
-	public class DomainHandlerExtensions
-	{
-		public DomainHandlerExtensions()
-		{
-			// public static void AddDomainHandlers(this IServiceCollection service)
-			// {
-			// 	
-			// }
-		}
-	}
+    public static class DomainHandlerExtensions
+    {
+        public static void AddDomainHandlers(this IServiceCollection services)
+        {
+            services.AddTransient<ISignUpHandler, SignUpHandler>();
+        }
+    }
 }
 
