@@ -3,6 +3,6 @@ using Microsoft.AspNetCore.Identity;
 namespace TwojUrlop.Common.Models.Entities;
 public class Role : IdentityRole<int>
 {
-    public ICollection<UserRole> UserRoles { get; set; }
-    public IEnumerable<User> Users { get; set; }
+    public virtual ICollection<UserRole> UserRoles { get; set; } = default!;
+    public virtual IEnumerable<User> Users { get; set; } = default!;
 }
