@@ -1,5 +1,4 @@
 namespace TwojUrlop.Common.Models.Interfaces;
 
 public interface ITwojUrlopDbContext
-{
-}
+{ }
