@@ -2,5 +2,5 @@ namespace TwojUrlop.Common.Models.Settings;
 
 public class ConnectionStringSettings
 {
-    public string? TwojUrlopDatabaseConnection { get; }
+    public string? TwojUrlopDatabaseConnection { get; } = default!;
 }
