@@ -3,18 +3,13 @@ using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using TwojUrlop.Common.Models.Entities;
 
 namespace TwojUrlop.DataAcess.Configuration;
-
 public class UserConfiguration : IEntityTypeConfiguration<User>
 {
     public void Configure(EntityTypeBuilder<User> builder)
     {
         builder.ToTable("User");
-        builder.HasKey(x => x.Id);
 
-        builder.HasOne(x => x.Role)
-            .WithMany(x => x.Users)
-            .HasForeignKey(x => x.RoleId)
-            .OnDelete(DeleteBehavior.Restrict);
+        builder.HasKey(x => x.Id);
     }
-    
+
 }
