@@ -5,30 +5,29 @@ using Microsoft.OpenApi.Models;
 using TwojUrlop.Common.Constrants;
 
 namespace TwojUrlop.DependencyInjection.Extensions;
-
-    public static class SwagerConfigExtensions
+public static class SwagerConfigExtensions
+{
+    public static void AddSwagger(this IServiceCollection service, string projectName)
     {
-        public static void AddSwagger(this IServiceCollection service, string projectName)
-        {
 
-            service.AddSwaggerGen(c =>
+        service.AddSwaggerGen(c =>
+        {
+            c.SwaggerDoc("v1", new OpenApiInfo { Title = projectName, Version = "v1" });
+            c.AddSecurityDefinition(TokenNames.BEARER, new OpenApiSecurityScheme
             {
-                c.SwaggerDoc("v1", new OpenApiInfo { Title = projectName, Version = "v1" });
-                c.AddSecurityDefinition(TokenNames.BEARER, new OpenApiSecurityScheme
-                {
-                    Description = @"JWT Authorization header using the Bearer scheme.
+                Description = @"JWT Authorization header using the Bearer scheme.
                       Enter 'Bearer' and then your token in the text input below.
                       Example: 'Bearer 12345abcdef'",
-                    Name = "Authorization",
-                    In = ParameterLocation.Header,
-                    Type = SecuritySchemeType.ApiKey,
-                    Scheme = TokenNames.BEARER
-                });
+                Name = "Authorization",
+                In = ParameterLocation.Header,
+                Type = SecuritySchemeType.ApiKey,
+                Scheme = TokenNames.BEARER
+            });
 
-                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-                c.AddSecurityRequirement(new OpenApiSecurityRequirement()
-                {
+            var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
+            var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
+            c.AddSecurityRequirement(new OpenApiSecurityRequirement()
+            {
                     {
                         new OpenApiSecurityScheme
                         {
@@ -43,18 +42,18 @@ namespace TwojUrlop.DependencyInjection.Extensions;
                         },
                         new List<string>()
                     }
-                });
             });
-        }
-
-        public static void UseSwagger(this IApplicationBuilder app)
-        {
-            SwaggerBuilderExtensions.UseSwagger(app);
-
-            app.UseSwaggerUI(c =>
-            {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "API");
-                c.RoutePrefix = string.Empty;
-            });
-        }
+        });
     }
+
+    public static void UseSwagger(this IApplicationBuilder app)
+    {
+        SwaggerBuilderExtensions.UseSwagger(app);
+
+        app.UseSwaggerUI(c =>
+        {
+            c.SwaggerEndpoint("/swagger/v1/swagger.json", "API");
+            c.RoutePrefix = string.Empty;
+        });
+    }
+}
